Extract QPSG and QPS tests into common
