+++ nncf/torch/quantization/algo.py
@@ -47,6 +47,7 @@ from nncf.common.initialization.batchnorm_adaptation import BatchnormAdaptationA
 from nncf.common.insertion_point_graph import InsertionPointGraph
 from nncf.common.insertion_point_graph import InsertionPointGraphNodeType
 from nncf.common.quantization.config_assignment import assign_qconfig_lists_to_modules
+from nncf.common.quantization.quantizer_setup import DEFAULT_QUANTIZER_CONFIG
 from nncf.common.quantization.quantizer_setup import MultiConfigQuantizerSetup
 from nncf.common.quantization.quantizer_setup import QuantizationPointId
 from nncf.common.quantization.quantizer_setup import QuantizerSetupBase
@@ -161,11 +162,6 @@ class QuantizerBuilderStateVersion(IntEnum):
 
 
 class QuantizerSetupGeneratorBase:
-    DEFAULT_QUANTIZER_CONFIG = QuantizerConfig(num_bits=8,
-                                               mode=QuantizationMode.SYMMETRIC,
-                                               signedness_to_force=None,
-                                               per_channel=False)
-
     def __init__(self, quant_config: Dict,
                  target_model: NNCFNetwork,
                  precision_init_type: str = None,
@@ -218,7 +214,7 @@ class QuantizerSetupGeneratorBase:
         self._target_scopes_per_group[quantizer_group] = params_dict_from_config.get('target_scopes')
 
     def _get_default_qconfig(self, constraints: QuantizationConstraints = None):
-        qconfig = deepcopy(self.DEFAULT_QUANTIZER_CONFIG)
+        qconfig = deepcopy(DEFAULT_QUANTIZER_CONFIG)
         if constraints is not None:
             qconfig = constraints.apply_constraints_to(qconfig)
         return qconfig
