+++ tests/common/quantization/metatypes.py
@@ -1,0 +1,88 @@
+from typing import List
+
+from nncf.common.graph import OperatorMetatype
+from nncf.common.graph.operator_metatypes import OperatorMetatypeRegistry
+from nncf.common.graph.operator_metatypes import UnknownMetatype
+from nncf.common.quantization.quantizer_propagation.structs import QuantizationTrait
+
+METATYPES_FOR_TEST = OperatorMetatypeRegistry('TEST_METATYPES')
+
+
+class TestMetatype(OperatorMetatype):
+    name = None
+
+    @classmethod
+    def get_all_aliases(cls) -> List[str]:
+        assert cls.name is not None
+        return [cls.name]
+
+
+@METATYPES_FOR_TEST.register()
+class BatchNormTestMetatype(TestMetatype):
+    name = 'batch_norm'
+
+
+@METATYPES_FOR_TEST.register()
+class Conv2dTestMetatype(TestMetatype):
+    name = 'conv2d'
+
+
+@METATYPES_FOR_TEST.register()
+class MatMulTestMetatype(TestMetatype):
+    name = 'matmul'
+
+
+@METATYPES_FOR_TEST.register()
+class MaxPool2dTestMetatype(TestMetatype):
+    name = 'max_pool2d'
+
+
+@METATYPES_FOR_TEST.register()
+class GeluTestMetatype(TestMetatype):
+    name = 'gelu'
+
+
+@METATYPES_FOR_TEST.register()
+class DropoutTestMetatype(TestMetatype):
+    name = 'dropout'
+
+
+@METATYPES_FOR_TEST.register()
+class MinTestMetatype(TestMetatype):
+    name = 'min'
+
+
+@METATYPES_FOR_TEST.register()
+class SoftmaxTestMetatype(TestMetatype):
+    name = 'softmax'
+
+
+@METATYPES_FOR_TEST.register()
+class CatTestMetatype(TestMetatype):
+    name = 'cat'
+
+
+@METATYPES_FOR_TEST.register()
+class LinearTestMetatype(TestMetatype):
+    name = 'linear'
+
+
+DEFAULT_TEST_QUANT_TRAIT_MAP = {
+    QuantizationTrait.INPUTS_QUANTIZABLE: [
+        BatchNormTestMetatype,
+        Conv2dTestMetatype,
+        MatMulTestMetatype,
+        GeluTestMetatype,
+        LinearTestMetatype
+    ],
+    QuantizationTrait.NON_QUANTIZABLE: [
+        MaxPool2dTestMetatype,
+        DropoutTestMetatype,
+        MinTestMetatype,
+        SoftmaxTestMetatype,
+        UnknownMetatype
+    ],
+    QuantizationTrait.CONCAT: [
+        CatTestMetatype
+    ],
+}
