+++ tests/common/quantization/test_quantizer_propagation_graph.py
@@ -41,12 +41,12 @@ from nncf.common.quantization.quantizer_setup import WeightQuantizationInsertion
 from nncf.common.quantization.structs import QuantizationMode
 from nncf.common.quantization.structs import QuantizerConfig
 from nncf.common.quantization.structs import UnifiedScaleType
-from nncf.torch.graph.operator_metatypes import PTCatMetatype
-from tests.torch.test_nncf_network import get_ip_graph_for_test
-from tests.torch.test_nncf_network import get_mock_nncf_node_attrs
-from tests.torch.test_nncf_network import get_nncf_graph_from_mock_nx_graph
-from tests.torch.test_nncf_network import get_two_branch_mock_model_graph
-from tests.torch.test_nncf_network import mark_input_ports_lexicographically_based_on_input_node_key
+from tests.common.quantization.metatypes import CatTestMetatype
+from tests.common.quantization.mock_graphs import get_ip_graph_for_test
+from tests.common.quantization.mock_graphs import get_mock_nncf_node_attrs
+from tests.common.quantization.mock_graphs import get_nncf_graph_from_mock_nx_graph
+from tests.common.quantization.mock_graphs import get_two_branch_mock_model_graph
+from tests.common.quantization.mock_graphs import mark_input_ports_lexicographically_based_on_input_node_key
 
 
 #pylint:disable=too-many-lines
@@ -72,7 +72,7 @@ class TestQuantizerPropagationStateGraph:
         ip_graph = get_ip_graph_for_test(get_two_branch_mock_model_graph())
         qpsg = QPSG(ip_graph)
 
-        qpsg.nodes['5 /F_0'][QPSG.OPERATOR_METATYPE_NODE_ATTR] = PTCatMetatype
+        qpsg.nodes['5 /F_0'][QPSG.OPERATOR_METATYPE_NODE_ATTR] = CatTestMetatype
         qpsg.skip_check = False
         yield qpsg
         if not qpsg.skip_check:
@@ -250,7 +250,7 @@ class TestQuantizerPropagationStateGraph:
         test_groups_vs_paths = \
             mock_qp_graph.get_paths_to_immediately_dominating_insertion_points_grouped_by_unified_scales(
                 start_node_key,
-                {PTCatMetatype})
+                {CatTestMetatype})
 
         def get_cat_path_list(path_list: List[List[Tuple[str, str]]]):
             str_paths = [[str(edge[0]) + ' -> ' + str(edge[1]) for edge in path] for path in path_list]
