+++ tests/torch/quantization/test_hawq_precision_init.py
@@ -68,12 +68,12 @@ from tests.torch.helpers import BasicConvTestModel
 from tests.torch.helpers import create_compressed_model_and_algo_for_test
 from tests.torch.helpers import create_conv
 from tests.torch.helpers import register_bn_adaptation_init_args
-from tests.torch.quantization.test_quantization_helpers import compare_multi_gpu_dump
-from tests.torch.quantization.test_quantization_helpers import create_rank_dataloader
-from tests.torch.quantization.test_quantization_helpers import distributed_init_test_default
-from tests.torch.quantization.test_quantization_helpers import get_quantization_config_without_range_init
-from tests.torch.quantization.test_quantization_helpers import get_squeezenet_quantization_config
-from tests.torch.quantization.test_quantization_helpers import post_compression_test_distr_init
+from tests.torch.quantization.quantization_helpers import compare_multi_gpu_dump
+from tests.torch.quantization.quantization_helpers import create_rank_dataloader
+from tests.torch.quantization.quantization_helpers import distributed_init_test_default
+from tests.torch.quantization.quantization_helpers import get_quantization_config_without_range_init
+from tests.torch.quantization.quantization_helpers import get_squeezenet_quantization_config
+from tests.torch.quantization.quantization_helpers import post_compression_test_distr_init
 # pylint:disable=unused-import
 from tests.torch.modules.test_rnn import _seed
 from tests.torch.test_compressed_graph import get_full_path_to_the_graph
