+++ tests/torch/test_knowledge_distillation.py
@@ -26,8 +26,8 @@ from tests.torch.helpers import create_compressed_model_and_algo_for_test
 from tests.torch.helpers import create_ones_mock_dataloader
 from tests.torch.helpers import fill_params_of_model_by_normal
 from tests.torch.sparsity.magnitude.test_helpers import get_basic_magnitude_sparsity_config
-from tests.torch.quantization.test_quantization_helpers import create_rank_dataloader, post_compression_test_distr_init
-from tests.torch.quantization.test_quantization_helpers import distributed_init_test_default
+from tests.torch.quantization.quantization_helpers import create_rank_dataloader, post_compression_test_distr_init
+from tests.torch.quantization.quantization_helpers import distributed_init_test_default
 
 import torch
 from torch import nn
